requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.29.0
pydantic[email]>=2.6.4
pydantic-settings>=2.2.1
google-genai>=0.3.0
python-dotenv>=1.0.1

# Database
sqlalchemy[asyncio]>=2.0.30
aiosqlite>=0.19.0
alembic>=1.13.1

# Authentication
python-jose[cryptography]>=3.3.0
bcrypt>=4.0.0
python-multipart>=0.0.9

# Rate limiting
slowapi>=0.1.9

# Logging
loguru>=0.7.2

# Testing
pytest>=8.1.1
pytest-asyncio>=0.23.5
httpx>=0.27.0

# Code Quality
ruff>=0.4.0
mypy>=1.10.0
